requirements.txt
appdirs==1.4.4
astroid==2.5.6
atomicwrites==1.4.0
attrs==21.2.0
black==21.5b1
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==8.0.1
coverage==4.5.1
coveralls==1.5.0
cryptography==3.4.7
defusedxml==0.7.1
docopt==0.6.2
Flask==1.0.2
gunicorn==19.9.0
<<<<<<< Updated upstream
coverage==4.5.2
=======
httmock==1.2.6
idna==2.7
isodate==0.6.0
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
lxml==4.6.3
MarkupSafe==2.0.1
mccabe==0.6.1
more-itertools==8.7.0
mypy-extensions==0.4.3
>>>>>>> Stashed changes
nose==1.3.7
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.10.1
pylint==2.1.1
pyOpenSSL==20.0.1
pytest==3.8.0
pytest-cov==2.6.0
<<<<<<< Updated upstream
Werkzeug==0.14.1
psycopg2==2.7.5
flask-jwt-extended==3.13.1
validate_email==1.3
flasgger==0.9.1
yapf==0.24.0
flask-cors==3.0.7
=======
python-decouple==3.4
pytz==2021.1
regex==2021.4.4
requests==2.19.1
requests-file==1.5.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
suds-community==0.8.5
toml==0.10.2
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.12.1
zeep==3.4.0
>>>>>>> Stashed changes
